// src/ejemplos/interfaces/PokemonResponse.ts
    export interface Ability2 {
        name: string;
        url: string;
    }

    export interface Ability {
        ability: Ability2;
        is_hidden: boolean;
        slot: number;
    }

    export interface Form {
        name: string;
        url: string;
    }

    export interface Version {
        name: string;
        url: string;
    }

    export interface GameIndice {
        game_index: number;
        version: Version;
    }

    export interface Move2 {
        name: string;
        url: string;
    }

    export interface MoveLearnMethod {
        name: string;
        url: string;
    }

    export interface VersionGroup {
        name: string;
        url: string;
    }

    export interface VersionGroupDetail {
        level_learned_at: number;
        move_learn_method: MoveLearnMethod;
        version_group: VersionGroup;
    }

    export interface Move {
        move: Move2;
        version_group_details: VersionGroupDetail[];
    }

    export interface Species {
        name: string;
        url: string;
    }

    export interface DreamWorld {
        front_default: string;
        front_female?: any;
    }

    export interface Home {
        front_default: string;
        front_female?: any;
        front_shiny: string;
        front_shiny_female?: any;
    }

    export interface OfficialArtwork {
        front_default: string;
    }

    export interface Other {
        dream_world: DreamWorld;
        home: Home;
        "official-artwork": OfficialArtwork;
    }
// ...
    export interface Versions {
        "generation-i": GenerationI;
        "generation-ii": GenerationIi;
        "generation-iii": GenerationIii;
        "generation-iv": GenerationIv;
        "generation-v": GenerationV;
        "generation-vi": GenerationVi;
        "generation-vii": GenerationVii;
        "generation-viii": GenerationViii;
    }

    export interface Sprites {
        back_default: string;
        back_female?: any;
        back_shiny: string;
        back_shiny_female?: any;
        front_default: string;
        front_female?: any;
        front_shiny: string;
        front_shiny_female?: any;
        other: Other;
        versions: Versions;
    }

    export interface Stat2 {
        name: string;
        url: string;
    }

    export interface Stat {
        base_stat: number;
        effort: number;
        stat: Stat2;
    }

    export interface Type2 {
        name: string;
        url: string;
    }

    export interface Type {
        slot: number;
        type: Type2;
    }

    export interface Pokemon {
        abilities: Ability[];
        base_experience: number;
        forms: Form[];
        game_indices: GameIndice[];
        height: number;
        held_items: any[];
        id: number;
        is_default: boolean;
        location_area_encounters: string;
        moves: Move[];
        name: string;
        order: number;
        past_types: any[];
        species: Species;
        sprites: Sprites;
        stats: Stat[];
        types: Type[];
        weight: number;
    }